requirements.txt
Flask>=3.0
python-dotenv
tzdata
